+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.7] - 2026-05-02
+
+- add search to posts @fdaei
+- add template sms @fdaei
+
+### Fixed
+
 ## [1.3.6] - 2026-03-26
 
 ### Added
@@ -168,7 +175,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - initial release 🎉​🎊​.
 
-[unreleased]: https://github.com/wenex-org/platform-sdk/compare/1.3.6...HEAD
+[unreleased]: https://github.com/wenex-org/platform-sdk/compare/1.3.7...HEAD
+[1.3.7]: https://github.com/wenex-org/platform-sdk/compare/1.3.6...1.3.7
 [1.3.6]: https://github.com/wenex-org/platform-sdk/compare/1.3.5...1.3.6
 [1.3.5]: https://github.com/wenex-org/platform-sdk/compare/1.3.4...1.3.5
 [1.3.4]: https://github.com/wenex-org/platform-sdk/compare/1.3.3...1.3.4